+++ src/forms/findOrCreateEntityForm.tsx
@@ -154,7 +154,7 @@ export class FindOrCreateEntityForm extends React.Component<FindOrCreateEntityFo
                         title={t.text('visual_authoring.dialog.apply.title')}>
                         {t.text('visual_authoring.dialog.apply.label')}
                     </button>
-                    <button className='reactodia-btn reactodia-btn-default'
+                    <button className='reactodia-btn reactodia-btn-secondary'
                         onClick={this.props.onCancel}
                         title={t.text('visual_authoring.dialog.cancel.title')}>
                         {t.text('visual_authoring.dialog.cancel.label')}